[soil] Begin soil sensing subproject
Support the Adafruit Stemma capacitive sensor.  This one was bad to wire
up - it came with a connector that didn't match breadboard pitch, so I
ended up removing that and soldering new leads.  It supports a
temperature sensor that is, according to the docs, is "not high
precision, maybe good to + or - 2 degrees Celsius" - i.e., it's shit.

Signed-off-by: Robbie Harwood <rharwood@pm.me>